GPST-3 5 Exceptions working test profile entity changes: long to int, repo same

+++ src/main/kotlin/no/nowak/gpstracker/core/address/AddressRepository.kt
@@ -2,8 +2,8 @@ package no.nowak.gpstracker.core.address
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface AddressRepository : JpaRepository<Address, Long> {
+interface AddressRepository : JpaRepository<Address, Int> {
 }
 
-interface CountryRepository : JpaRepository<Country, Long> {
+interface CountryRepository : JpaRepository<Country, Int> {
 }
