+++ src/main/kotlin/no/nowak/gpstracker/core/device/DeviceRepository.kt
@@ -2,5 +2,5 @@ package no.nowak.gpstracker.core.device
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface DeviceRepository : JpaRepository<Device, Long> {
+interface DeviceRepository : JpaRepository<Device, Int> {
 }
